% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synth_corpus.R
\name{gen_synth_corpus_multi_h}
\alias{gen_synth_corpus_multi_h}
\title{Generates a Corpus with Multiple \eqn{(\alpha, \eta)}s}
\usage{
gen_synth_corpus_multi_h(K, V, J, collection.size, doc.size, alpha.vec, eta.vec)
}
\arguments{
\item{K}{number of topics}

\item{V}{vocabulary size}

\item{J}{number of collections}

\item{collection.size}{number of documents in each collection (a list of J
elements)}

\item{doc.size}{number of words in each document}

\item{alpha.vec}{a set of values for document-level Dirichlet sampling}

\item{eta.vec}{a set of values for topic Diriclet sampling}
}
\value{
a list of generated corpus and their statistics
}
\description{
Generates documents using the LDA generative process based on a set of
\eqn{(\alpha, \eta)} Configurations
}
\details{
Last modified on: April 15, 2016
}
\examples{
## Generates documents with given parameters

J                  <- 2
K                  <- 4
V                  <- 20
alpha.vec          <- c(.2, 2)
eta.vec            <- c(.5, 2)
doc.size           <- 80
collection.size    <- c(40, 40) # number of documents in each collection
ds.name            <- paste("synth-J", J, "-K", K, "-V", V, sep = "")

ds                 <- gen_synth_corpus_multi_h(K, V, J, collection.size, doc.size, alpha.vec, eta.vec)

}
\seealso{
Other corpus: \code{\link{calc_doc_cos}},
  \code{\link{calc_doc_tf}},
  \code{\link{gen_synth_corpus2}},
  \code{\link{gen_synth_corpus_multi_alpha}}
}

