% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lda_vem}
\alias{lda_vem}
\title{LDA: Variational EM}
\usage{
lda_vem(num_topics, vocab_size, docs_tf, alpha_h, eta_h, vi_max_iter,
  em_max_iter, vi_conv_thresh, em_conv_thresh, estimate_alpha, estimate_eta,
  verbose)
}
\arguments{
\item{num_topics}{Number of topics in the corpus}

\item{vocab_size}{Vocabulary size}

\item{docs_tf}{A list of corpus documents read from the Blei corpus using 
\code{\link{read_docs}} (term indices starts with 0)}

\item{alpha_h}{Hyperparameter for \eqn{\theta} sampling}

\item{eta_h}{Smoothing parameter for the \eqn{\beta} matrix}

\item{vi_max_iter}{Maximum number of iterations for variational inference}

\item{em_max_iter}{Maximum number of iterations for variational EM}

\item{estimate_alpha}{If true, run hyperparameter \eqn{\alpha} optimization}

\item{estimate_eta}{If true, run hyperparameter \eqn{\eta} optimization}

\item{verbose}{from {0, 1, 2}}

\item{vi_conv_threshold}{Convergence threshold for the document variational
inference loop}

\item{em_conv_threshold}{Convergence threshold for the variational EM loop}
}
\value{
TBA
}
\description{
This implements the variational inference algorithm for the LDA (full 
Bayesian) model. This includes optimization routines for both \eqn{\alpha} 
and \eqn{\eta} hyperparameters.
}
\details{
References: 
  * Latent Dirichlet Allocation. D. Blei, A. Ng, M.I. Jordan (2003)
}
\note{
Created on April 26, 2016
}

